requirements.txt
flask
flask-cors
tflite-runtime
Pillow
numpy